[proliferate] split session selection actions
