+++ desktop/src/hooks/sessions/workflows/session-selection-runtime.ts
@@ -1,0 +1,53 @@
+import {
+  getLatencyFlowRequestHeaders,
+} from "@/lib/infra/measurement/latency-flow";
+import {
+  getMeasurementRequestOptions,
+  type MeasurementOperationId,
+} from "@/lib/infra/measurement/debug-measurement";
+import { bootstrapHarnessRuntime } from "@/lib/access/anyharness/runtime-bootstrap";
+import { fetchWorkspaceSessionSummaries } from "@/lib/workflows/sessions/session-runtime";
+import type { WorkspaceSession } from "@/hooks/access/anyharness/sessions/use-workspace-session-cache";
+import { useHarnessConnectionStore } from "@/stores/sessions/harness-connection-store";
+
+export function buildLatencyRequestOptions(latencyFlowId?: string | null) {
+  const headers = getLatencyFlowRequestHeaders(latencyFlowId);
+  return headers ? { headers } : undefined;
+}
+
+export async function ensureRuntimeReadyForSessions(): Promise<string> {
+  const state = useHarnessConnectionStore.getState();
+  if (state.connectionState !== "healthy" || state.runtimeUrl.trim().length === 0) {
+    await bootstrapHarnessRuntime();
+  }
+
+  const readyState = useHarnessConnectionStore.getState();
+  if (readyState.connectionState !== "healthy" || readyState.runtimeUrl.trim().length === 0) {
+    throw new Error(readyState.error || "AnyHarness runtime is still starting. Try again.");
+  }
+
+  return readyState.runtimeUrl;
+}
+
+export async function fetchWorkspaceSessions(
+  runtimeUrl: string,
+  workspaceId: string,
+  options?: {
+    requestHeaders?: HeadersInit;
+    measurementOperationId?: MeasurementOperationId | null;
+  },
+): Promise<WorkspaceSession[]> {
+  const sessions = await fetchWorkspaceSessionSummaries(
+    runtimeUrl,
+    workspaceId,
+    getMeasurementRequestOptions({
+      operationId: options?.measurementOperationId,
+      category: "session.list",
+      headers: options?.requestHeaders,
+    }),
+  );
+  return sessions.map((session) => ({
+    ...session,
+    workspaceId,
+  }));
+}
