+++ desktop/src/hooks/workspaces/selection/types.ts
@@ -1,6 +1,6 @@
 import type { AnyHarnessResolvedConnection } from "@anyharness/sdk-react";
 import type { Workspace } from "@anyharness/sdk";
-import type { WorkspaceSession } from "@/hooks/sessions/use-session-selection-actions";
+import type { WorkspaceSession } from "@/hooks/access/anyharness/sessions/use-workspace-session-cache";
 import type { LogicalWorkspace } from "@/lib/domain/workspaces/cloud/logical-workspaces";
 
 export interface WorkspaceSelectionOptions {
